rename data.rs to decoded_array.rs
