+++ src/lib.rs
@@ -4,7 +4,7 @@ mod array;
 mod array_bytes;
 mod chunks;
 mod codec;
-mod data;
+mod decoded_array;
 mod dtype;
 mod error;
 mod exceptions;
@@ -20,7 +20,7 @@ use crate::array::{PyArray, PyAsyncArray};
 use crate::array_bytes::PyArrayBytes;
 use crate::chunks::PyChunkGrid;
 use crate::codec::register_codec_module;
-use crate::data::{PyMaskedTensor, PyMaskedVariableArray, PyTensor, PyVariableArray};
+use crate::decoded_array::{PyMaskedTensor, PyMaskedVariableArray, PyTensor, PyVariableArray};
 use crate::dtype::PyDataType;
 use crate::exceptions::register_exceptions_module;
 use crate::fill_value::PyFillValue;
